pyproject.toml
[tool.poetry]
name = "fastapi-blog"
version = "0.1.0"
description = ""
authors = ["jamshidyerzakov <51641366+jamik63@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8.2"
fastapi = "^0.61.1"
uvicorn = "^0.11.8"
tortoise-orm = "^0.16.14"
asyncpg = "^0.21.0"
aiosqlite = "^0.15.0"
aiomysql = "^0.0.20"
SQLAlchemy = "^1.3.19"
alembic = "^1.4.3"
psycopg2 = "^2.8.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
